+++ tests/test_construct/test_shape.py
@@ -1,7 +1,10 @@
 import unittest
 import numpy as np
 
+from classy_blocks.construct.flat.face import Face
 from classy_blocks.construct.shapes.elbow import Elbow
+from classy_blocks.construct.shapes.rings import RevolvedRing
+from classy_blocks.construct.shapes.sphere import Hemisphere
 from classy_blocks.util import functions as f
 
 
@@ -52,6 +55,40 @@ class ElbowTests(unittest.TestCase):
         self.assertAlmostEqual(self.elbow.sketch_mid.radius, (self.radius_1 + self.radius_2) / 2)
 
 
+class RevolvedRingTests(unittest.TestCase):
+    """RevolvedRing creation and manipulation"""
+
+    def setUp(self):
+        self.face = Face([[0, 1, 0], [1, 1, 0], [1, 2, 0], [0, 2, 0]])
+        self.axis_point_1 = [0, 0, 0]
+        self.axis_point_2 = [1, 0, 0]
+        self.n_segments = 8
+
+    @property
+    def ring(self) -> RevolvedRing:
+        """The test subject"""
+        return RevolvedRing(self.axis_point_1, self.axis_point_2, self.face, self.n_segments)
+
+    def test_create(self):
+        """Create a revolved ring"""
+        _ = self.ring
+
+    def test_set_inner_patch(self):
+        """Inner faces of the ring"""
+        ring = self.ring
+        ring.set_inner_patch("inner")
+
+        for operation in ring.operations:
+            self.assertEqual(operation.patch_names["front"], "inner")
+
+
 class SphereTests(unittest.TestCase):
-    # TODO: points, transforms
-    pass
+    def test_sphere_radii(self):
+        """Check that all radii, defined by outer points, are as specified"""
+        center = f.vector(1, 1, 1)
+        sphere = Hemisphere(center, [2, 2, 1], [0, 0, 1])
+        radius = 2**0.5
+
+        for loft in sphere.shell:
+            for point in loft.get_face("right").point_array:
+                self.assertAlmostEqual(f.norm(point - center), radius)